Removed lib folder
